+++ CHANGELOG.md
@@ -14,6 +14,12 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.1.2.25] - 2021-02-25
+
+### Added
+- translation for tables
+- translation for order table
+
 ## [2021.1.2.23] - 2021-02-23
 
 ### Added